Splitting out files
* Moving class types into separate files
* Convenience method for loading types
